+++ CHANGELOG.md
@@ -13,6 +13,20 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 - Fix all Subnet Proxy issues (as if)
 
 
+## [0.6.8-beta] - 2026-??-??
+### Added
+- Prevent redstone updates from triggering unnecessary capability scans on adjacent tiles for Interfaces, if the actual tiles have not changed.
+- Add right-click to insert compatible upgrades directly into interfaces/subnet proxies.
+- Add Ctrl-A/Ctrl-C/Ctrl-V support to the Max Slot Size text field.
+- Add a config to change the default +/- 1/10/100/1000 offset buttons to different values.
+- Add a config to use fixed values instead of offsets for the Max Slot Size buttons.
+- Allow Interfaces and Subnet Proxies to restore upgrades from the player's inventory when restoring from a memory card, if the upgrades are missing from the interface/proxy itself. This allows to copy filters and upgrades from one interface to another without having to manually insert the cards. Pull/Push cards have their settings restored on top of being inserted, meaning you do not need to have the specific configuration in the player's inventory, just the card itself.
+
+### Fixed
+- Fix the Clear Filters button not applying in I/O and Universal Interfaces.
+- Fix Subnet Proxy not properly initializing when adding the first filters (resulting in nothing being forwarded until reloaded).
+
+
 ## [0.6.7-beta2] - 2026-08-11
 ### Added
 - Add a manual refresh button to interface GUIs so the current tab can immediately push/pull all content to the network, including Pull/Push Card operations.